refactor(core): split event delivery and failed display loop specs
