+++ packages/agent-core/test/Agent/LoopSpec.hs
@@ -5,6 +5,9 @@ import qualified Agent.Json.Decode as Json
 import Agent.Cancel (newCancelFlag, requestCancel)
 import Agent.Error (ApiError(..))
 import Agent.Loop
+import Agent.Loop.Fixtures
+import qualified Agent.Loop.EventDeliverySpec as EventDelivery
+import qualified Agent.Loop.FailedDisplaySpec as FailedDisplay
 import Agent.Responses.Types
     ( FunctionCallOutput(..)
     , MessageContent(..)
@@ -13,11 +16,9 @@ import Agent.Responses.Types
     , ResponseItem(..)
     , ResponseMessage(..)
     , ResponseRole(..)
-    , TaggedObject(..)
     )
-import Agent.Telemetry (TurnTelemetry(..))
-import Agent.ToolArgs (objectArgs, reqText)
 import Agent.ToolDispatch
+import Agent.Telemetry (TurnTelemetry(..))
 import Agent.Tools.Scheduling
     ( ToolAccess(..)
     , ToolResource(..)
@@ -26,14 +27,10 @@ import Agent.Tools.Scheduling
     , schedulingPlansConflict
     )
 import Agent.Tools.Types
-    ( AppTool(..)
-    , ApprovalRule(..)
+    ( ApprovalRule(..)
     , ToolExecutionPolicy(..)
-    , ToolRegistry
     , jsonAppToolWithExecution
-    , mkToolRegistry
     , toolExecutionPolicyFor
-    , withAsyncToolCalls
     , withToolResourceClaims
     )
 import Codec.Picture
@@ -85,17 +82,6 @@ import Data.Word (Word32)
 import System.Timeout (timeout)
 import System.OsPath (unsafeEncodeUtf)
 import Test.Hspec
-
-emptyTestTelemetry :: TurnTelemetry
-emptyTestTelemetry = TurnTelemetry
-    { telemetryDurationMs = Nothing
-    , telemetryApiDurationMs = Nothing
-    , telemetryCostUsd = Nothing
-    , telemetryStopReason = Nothing
-    , telemetryProviderTurns = Nothing
-    , telemetryModels = mempty
-    , telemetryStructuredOutput = Nothing
-    }
 
 spec :: Spec
 spec = describe "runLoop" do
@@ -570,396 +556,7 @@ spec = describe "runLoop" do
         seen <- readIORef submissions
         seen `shouldBe` [(Nothing, inputs)]
 
-    it "serializes loopOnEvent across parallel tool calls" do
-        inFlight <- newIORef (0 :: Int)
-        maxInFlight <- newIORef (0 :: Int)
-        submissions <- newIORef []
-        backend <- scriptedBackend submissions
-            [ Right $ emptyTurnOutput "resp-1"
-                [ functionToolCall "c1" "a" "{}"
-                , functionToolCall "c2" "b" "{}"
-                ]
-                Nothing
-            , Right $ emptyTurnOutput "resp-2" [] (Just "ok")
-            ]
-        let onEvent _ = do
-                now <- atomicModifyIORef' inFlight \n -> (n + 1, n + 1)
-                atomicModifyIORef' maxInFlight \seen -> (max seen now, ())
-                threadDelay 30000
-                atomicModifyIORef' inFlight \n -> (n - 1, ())
-            handlers =
-                [ noArgsTool "a" (pure (Right "ok"))
-                , noArgsTool "b" (pure (Right "ok"))
-                ]
-        config0 <- testConfig backend
-        let config = config0
-                { loopTools = registryFromHandlers handlers
-                , loopOnEvent = onEvent
-                }
-        result <- runLoop config Nothing "go"
-        result `shouldBe` Right LoopResult
-            { finalResponseId = "resp-2"
-            , finalText = Just "ok"
-            , turnsUsed = 2
-            , tokenUsage = emptyTokenUsage
-            }
-        readIORef maxInFlight `shouldReturn` 1
-
-    it "delivers events off the backend thread and flushes before returning" do
-        sinkStarted <- newEmptyMVar
-        releaseSink <- newEmptyMVar
-        backendEntered <- newEmptyMVar
-        let backend = Backend \_state _prev _inputs _onEvent -> do
-                putMVar backendEntered ()
-                pure $ Right BackendResult
-                    { backendOutput =
-                        emptyTurnOutput "resp-1" [] (Just "done")
-                    , backendState = emptyBackendSnapshot
-                    }
-            onEvent = \case
-                TurnStarted -> do
-                    putMVar sinkStarted ()
-                    takeMVar releaseSink
-                _ -> pure ()
-        config0 <- testConfig backend
-        let config = config0 { loopOnEvent = onEvent }
-        withAsync (runLoop config Nothing "go") \running -> do
-            takeMVar sinkStarted
-            timeout 1000000 (takeMVar backendEntered)
-                `shouldReturn` Just ()
-            timeout 100000 (wait running)
-                `shouldReturn` Nothing
-            putMVar releaseSink ()
-            wait running `shouldReturn` Right LoopResult
-                { finalResponseId = "resp-1"
-                , finalText = Just "done"
-                , turnsUsed = 1
-                , tokenUsage = emptyTokenUsage
-                }
-
-    it "bounds queued events when the sink falls behind" do
-        sinkStarted <- newEmptyMVar
-        releaseSink <- newEmptyMVar
-        backendStarted <- newEmptyMVar
-        backendFinished <- newEmptyMVar
-        let backend = Backend \_state _prev _inputs onEvent -> do
-                putMVar backendStarted ()
-                mapM_ (const (onEvent (WarningRaised "x"))) [1 .. 300 :: Int]
-                putMVar backendFinished ()
-                pure $ Right BackendResult
-                    { backendOutput =
-                        emptyTurnOutput "resp-1" [] (Just "done")
-                    , backendState = emptyBackendSnapshot
-                    }
-            onEvent = \case
-                TurnStarted -> do
-                    putMVar sinkStarted ()
-                    takeMVar releaseSink
-                _ -> pure ()
-        config0 <- testConfig backend
-        let config = config0 { loopOnEvent = onEvent }
-        withAsync (runLoop config Nothing "go") \running -> do
-            takeMVar sinkStarted
-            takeMVar backendStarted
-            timeout 100000 (takeMVar backendFinished)
-                `shouldReturn` Nothing
-            putMVar releaseSink ()
-            timeout 3000000 (takeMVar backendFinished)
-                `shouldReturn` Just ()
-            wait running `shouldReturn` Right LoopResult
-                { finalResponseId = "resp-1"
-                , finalText = Just "done"
-                , turnsUsed = 1
-                , tokenUsage = emptyTokenUsage
-                }
-
-    it "wakes a producer blocked on a full queue when the sink fails" do
-        sinkStarted <- newEmptyMVar
-        releaseSink <- newEmptyMVar
-        backendStarted <- newEmptyMVar
-        backendFinished <- newEmptyMVar
-        let backend = Backend \_state _prev _inputs onEvent -> do
-                putMVar backendStarted ()
-                mapM_ (const (onEvent (WarningRaised "queued")))
-                    [1 .. 300 :: Int]
-                putMVar backendFinished ()
-                pure $ Right BackendResult
-                    { backendOutput =
-                        emptyTurnOutput "resp-1" [] (Just "done")
-                    , backendState = emptyBackendSnapshot
-                    }
-            onEvent = \case
-                TurnStarted -> do
-                    putMVar sinkStarted ()
-                    takeMVar releaseSink
-                    Exception.throwIO (userError "renderer exploded")
-                _ -> pure ()
-        config0 <- testConfig backend
-        let config = config0 { loopOnEvent = onEvent }
-        withAsync (runLoop config Nothing "go") \running -> do
-            takeMVar sinkStarted
-            takeMVar backendStarted
-            timeout 100000 (takeMVar backendFinished)
-                `shouldReturn` Nothing
-            putMVar releaseSink ()
-            timeout 1000000 (wait running)
-                `shouldReturn`
-                    Just
-                        (Left
-                            (LoopUnexpected
-                                "user error (renderer exploded)"))
-
-    it "coalesces adjacent deltas while preserving event boundaries" do
-        sinkStarted <- newEmptyMVar
-        releaseSink <- newEmptyMVar
-        backendFinished <- newEmptyMVar
-        events <- newIORef []
-        let backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (TextDelta "a")
-                onEvent (TextDelta "b")
-                onEvent (WarningRaised "boundary")
-                onEvent (ReasoningDelta "r1")
-                onEvent (ReasoningDelta "r2")
-                onEvent (TextDelta "c")
-                onEvent (TextDelta "d")
-                putMVar backendFinished ()
-                pure $ Right BackendResult
-                    { backendOutput =
-                        emptyTurnOutput "resp-1" [] (Just "done")
-                    , backendState = emptyBackendSnapshot
-                    }
-            onEvent event = do
-                modifyIORef' events (event :)
-                case event of
-                    TurnStarted -> do
-                        putMVar sinkStarted ()
-                        takeMVar releaseSink
-                    _ -> pure ()
-        config0 <- testConfig backend
-        let config = config0 { loopOnEvent = onEvent }
-        withAsync (runLoop config Nothing "go") \running -> do
-            takeMVar sinkStarted
-            takeMVar backendFinished
-            putMVar releaseSink ()
-            wait running `shouldReturn` Right LoopResult
-                { finalResponseId = "resp-1"
-                , finalText = Just "done"
-                , turnsUsed = 1
-                , tokenUsage = emptyTokenUsage
-                }
-        reverse <$> readIORef events `shouldReturn`
-            [ TurnStarted
-            , TextDelta "ab"
-            , WarningRaised "boundary"
-            , ReasoningDelta "r1r2"
-            , TextDelta "cd"
-            , TurnFinished (emptyTurnOutput "resp-1" [] (Just "done"))
-            ]
-
-    it "backpressures a coalesced text tail by logical payload bytes" do
-        sinkStarted <- newEmptyMVar
-        releaseSink <- newEmptyMVar
-        backendFinished <- newEmptyMVar
-        deliveredChars <- newIORef (0 :: Int)
-        let chunk = Text.replicate (1024 * 1024) "x"
-            backend = Backend \_state _prev _inputs onEvent -> do
-                -- The chunks cross the conservative 8 MiB logical-byte
-                -- budget while TurnStarted blocks the consumer, even though
-                -- they would otherwise occupy only one TBQueue node.
-                mapM_ (onEvent . TextDelta) [chunk, chunk, chunk]
-                putMVar backendFinished ()
-                pure $ Right BackendResult
-                    { backendOutput =
-                        emptyTurnOutput "resp-1" [] (Just "done")
-                    , backendState = emptyBackendSnapshot
-                    }
-            onEvent = \case
-                TurnStarted -> do
-                    putMVar sinkStarted ()
-                    takeMVar releaseSink
-                TextDelta text ->
-                    modifyIORef' deliveredChars (+ Text.length text)
-                _ -> pure ()
-        config0 <- testConfig backend
-        let config = config0 { loopOnEvent = onEvent }
-        withAsync (runLoop config Nothing "go") \running -> do
-            takeMVar sinkStarted
-            timeout 100000 (takeMVar backendFinished)
-                `shouldReturn` Nothing
-            putMVar releaseSink ()
-            timeout 1000000 (takeMVar backendFinished)
-                `shouldReturn` Just ()
-            wait running `shouldReturn` Right LoopResult
-                { finalResponseId = "resp-1"
-                , finalText = Just "done"
-                , turnsUsed = 1
-                , tokenUsage = emptyTokenUsage
-                }
-        readIORef deliveredChars
-            `shouldReturn` 3 * Text.length chunk
-
-    it "backpressures and coalesces provider-native child output" do
-        sinkStarted <- newEmptyMVar
-        releaseSink <- newEmptyMVar
-        backendFinished <- newEmptyMVar
-        deliveredChars <- newIORef (0 :: Int)
-        let chunk = Text.replicate (1024 * 1024) "x"
-            backend = Backend \_state _prev _inputs onEvent -> do
-                mapM_
-                    (onEvent . NativeAgentOutput "child")
-                    [chunk, chunk, chunk]
-                putMVar backendFinished ()
-                pure $ Right BackendResult
-                    { backendOutput =
-                        emptyTurnOutput "resp-1" [] (Just "done")
-                    , backendState = emptyBackendSnapshot
-                    }
-            onEvent = \case
-                TurnStarted -> do
-                    putMVar sinkStarted ()
-                    takeMVar releaseSink
-                NativeAgentOutput "child" output ->
-                    modifyIORef' deliveredChars (+ Text.length output)
-                _ -> pure ()
-        config0 <- testConfig backend
-        let config = config0 { loopOnEvent = onEvent }
-        withAsync (runLoop config Nothing "go") \running -> do
-            takeMVar sinkStarted
-            timeout 100000 (takeMVar backendFinished)
-                `shouldReturn` Nothing
-            putMVar releaseSink ()
-            timeout 1000000 (takeMVar backendFinished)
-                `shouldReturn` Just ()
-            wait running `shouldReturn` Right LoopResult
-                { finalResponseId = "resp-1"
-                , finalText = Just "done"
-                , turnsUsed = 1
-                , tokenUsage = emptyTokenUsage
-                }
-        readIORef deliveredChars
-            `shouldReturn` 3 * Text.length chunk
-
-    it "keeps only the latest adjacent tool-output snapshot per call" do
-        sinkStarted <- newEmptyMVar
-        releaseSink <- newEmptyMVar
-        backendFinished <- newEmptyMVar
-        events <- newIORef []
-        let backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (ToolOutputUpdated "c1" "a")
-                onEvent (ToolOutputUpdated "c1" "ab")
-                onEvent (ToolOutputUpdated "c2" "x")
-                onEvent (ToolOutputUpdated "c2" "xy")
-                onEvent (ToolOutputUpdated "c1" "abc")
-                putMVar backendFinished ()
-                pure $ Right BackendResult
-                    { backendOutput =
-                        emptyTurnOutput "resp-1" [] (Just "done")
-                    , backendState = emptyBackendSnapshot
-                    }
-            onEvent event = do
-                modifyIORef' events (event :)
-                case event of
-                    TurnStarted -> do
-                        putMVar sinkStarted ()
-                        takeMVar releaseSink
-                    _ -> pure ()
-        config0 <- testConfig backend
-        let config = config0 { loopOnEvent = onEvent }
-        withAsync (runLoop config Nothing "go") \running -> do
-            takeMVar sinkStarted
-            takeMVar backendFinished
-            putMVar releaseSink ()
-            wait running `shouldReturn` Right LoopResult
-                { finalResponseId = "resp-1"
-                , finalText = Just "done"
-                , turnsUsed = 1
-                , tokenUsage = emptyTokenUsage
-                }
-        reverse <$> readIORef events `shouldReturn`
-            [ TurnStarted
-            , ToolOutputUpdated "c1" "ab"
-            , ToolOutputUpdated "c2" "xy"
-            , ToolOutputUpdated "c1" "abc"
-            , TurnFinished (emptyTurnOutput "resp-1" [] (Just "done"))
-            ]
-
-    it "keeps only the latest adjacent tool-argument snapshot per call" do
-        sinkStarted <- newEmptyMVar
-        releaseSink <- newEmptyMVar
-        backendFinished <- newEmptyMVar
-        events <- newIORef []
-        let preview callId arguments =
-                ToolArgumentsUpdated
-                    (functionToolCall callId "apply_patch" arguments)
-            backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (preview "c1" "a")
-                onEvent (preview "c1" "ab")
-                onEvent (preview "c2" "x")
-                onEvent (preview "c2" "xy")
-                onEvent (preview "c1" "abc")
-                putMVar backendFinished ()
-                pure $ Right BackendResult
-                    { backendOutput =
-                        emptyTurnOutput "resp-1" [] (Just "done")
-                    , backendState = emptyBackendSnapshot
-                    }
-            onEvent event = do
-                modifyIORef' events (event :)
-                case event of
-                    TurnStarted -> do
-                        putMVar sinkStarted ()
-                        takeMVar releaseSink
-                    _ -> pure ()
-        config0 <- testConfig backend
-        let config = config0 { loopOnEvent = onEvent }
-        withAsync (runLoop config Nothing "go") \running -> do
-            takeMVar sinkStarted
-            takeMVar backendFinished
-            putMVar releaseSink ()
-            wait running `shouldReturn` Right LoopResult
-                { finalResponseId = "resp-1"
-                , finalText = Just "done"
-                , turnsUsed = 1
-                , tokenUsage = emptyTokenUsage
-                }
-        reverse <$> readIORef events `shouldReturn`
-            [ TurnStarted
-            , preview "c1" "ab"
-            , preview "c2" "xy"
-            , preview "c1" "abc"
-            , TurnFinished (emptyTurnOutput "resp-1" [] (Just "done"))
-            ]
-
-    it "bounds a single oversized live tool-output snapshot" do
-        delivered <- newIORef Nothing
-        let oversized =
-                Text.replicate (3 * 1024 * 1024) "x" <> "newest-tail"
-            backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (ToolOutputUpdated "large" oversized)
-                pure $ Right BackendResult
-                    { backendOutput =
-                        emptyTurnOutput "resp-1" [] (Just "done")
-                    , backendState = emptyBackendSnapshot
-                    }
-            onEvent = \case
-                ToolOutputUpdated "large" output ->
-                    writeIORef delivered (Just output)
-                _ -> pure ()
-        config0 <- testConfig backend
-        result <- runLoop config0 { loopOnEvent = onEvent } Nothing "go"
-        result `shouldBe` Right LoopResult
-            { finalResponseId = "resp-1"
-            , finalText = Just "done"
-            , turnsUsed = 1
-            , tokenUsage = emptyTokenUsage
-            }
-        readIORef delivered >>= \case
-            Nothing -> expectationFailure "missing tool-output update"
-            Just output -> do
-                Text.length output `shouldSatisfy` (<= 2 * 1024 * 1024)
-                output `shouldSatisfy`
-                    Text.isPrefixOf "[earlier tool output truncated]"
-                output `shouldSatisfy` Text.isSuffixOf "newest-tail"
+    EventDelivery.spec
 
     it "dispatches consecutive parallel-safe tool calls concurrently" do
         firstStarted <- newEmptyMVar
@@ -2007,156 +1604,7 @@ spec = describe "runLoop" do
         execution.executionResult
             `shouldBe` Left (LoopUnexpected "user error (renderer exploded)")
 
-    it "marks a transport failure after streamed output as interrupted" do
-        observed <- newIORef []
-        let backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (TextDelta "partial")
-                pure (Left (ConnectionError "down"))
-        config0 <- testConfig backend
-        let config = config0
-                { loopOnEvent = \event ->
-                    modifyIORef' observed (<> [event])
-                }
-        execution <-
-            runLoopInputsDetailed config Nothing [UserMessage "hello"]
-        execution.executionResult `shouldBe`
-            Left (LoopTransportAfterOutput (ConnectionError "down"))
-        execution.executionUncommittedDisplayEvents
-            `shouldBe` [TextDelta "partial"]
-        readIORef observed `shouldReturn`
-            [TurnStarted, TextDelta "partial", ResponseAttemptFailed]
-
-    it "coalesces retained text without crossing tool boundaries" do
-        let call =
-                functionToolCall "c1" "shell_command"
-                    "{\"command\":\"git status\"}"
-            backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (TextDelta "before")
-                onEvent (TextDelta " tool")
-                onEvent (ToolStarted call)
-                onEvent (TextDelta "after")
-                onEvent (TextDelta " tool")
-                pure (Left (ConnectionError "down"))
-        config <- testConfig backend
-        execution <-
-            runLoopInputsDetailed config Nothing [UserMessage "hello"]
-        execution.executionUncommittedDisplayEvents
-            `shouldBe`
-                [ TextDelta "before tool"
-                , ToolStarted call
-                , TextDelta "after tool"
-                ]
-
-    it "retains tool-only activity as display metadata on failure" do
-        let call =
-                functionToolCall "c1" "shell_command"
-                    "{\"command\":\"git status\"}"
-            backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (ToolStarted call)
-                onEvent (ToolOutputUpdated "c1" "still running")
-                pure (Left (ConnectionError "down"))
-        config <- testConfig backend
-        execution <-
-            runLoopInputsDetailed config Nothing [UserMessage "hello"]
-        execution.executionResult `shouldBe`
-            Left (LoopTransportAfterOutput (ConnectionError "down"))
-        execution.executionUncommittedDisplayEvents
-            `shouldBe`
-                [ ToolStarted call
-                , ToolOutputUpdated "c1" "still running"
-                ]
-
-    it "keeps earlier restarted attempts in display metadata" do
-        let backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (TextDelta "fir")
-                onEvent (TextDelta "st")
-                onEvent (ResponseRestarted "retrying")
-                onEvent (TextDelta "sec")
-                onEvent (TextDelta "ond")
-                pure (Left (ConnectionError "down"))
-        config <- testConfig backend
-        execution <-
-            runLoopInputsDetailed config Nothing [UserMessage "hello"]
-        execution.executionUncommittedDisplayEvents
-            `shouldBe`
-                [ TextDelta "first"
-                , ResponseRestarted "retrying"
-                , TextDelta "second"
-                ]
-        execution.executionUncommittedAssistantText
-            `shouldBe` Just "first\n\nsecond"
-
-    it "still marks failure after a later retry attempt is discarded" do
-        observed <- newIORef []
-        let backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (TextDelta "fir")
-                onEvent (TextDelta "st")
-                onEvent (ResponseRestarted "retrying")
-                onEvent (TextDelta "discard")
-                onEvent (TextDelta " me")
-                onEvent ResponseAttemptDiscarded
-                pure (Left (ConnectionError "down"))
-        config0 <- testConfig backend
-        let config = config0
-                { loopOnEvent = \event ->
-                    modifyIORef' observed (<> [event])
-                }
-        execution <-
-            runLoopInputsDetailed config Nothing [UserMessage "hello"]
-        execution.executionResult `shouldBe`
-            Left (LoopTransportAfterOutput (ConnectionError "down"))
-        execution.executionUncommittedDisplayEvents
-            `shouldBe`
-                [ TextDelta "first"
-                , ResponseRestarted "retrying"
-                ]
-        observedEvents <- readIORef observed
-        filter
-            (\case
-                TextDelta _ -> False
-                _ -> True)
-            observedEvents
-            `shouldBe`
-                [ TurnStarted
-                , ResponseRestarted "retrying"
-                , ResponseAttemptDiscarded
-                , ResponseAttemptFailed
-                ]
-
-    it "treats a transport failure after a discarded attempt as pre-output" do
-        let backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent (TextDelta "partial")
-                onEvent ResponseAttemptDiscarded
-                pure (Left (ConnectionError "down"))
-        config <- testConfig backend
-        result <- runLoop config Nothing "hello"
-        result `shouldBe` Left (LoopTransport (ConnectionError "down"))
-
-    it "bounds completed tool output retained after failure" do
-        let oversized =
-                Text.replicate (3 * 1024 * 1024) "x" <> "newest-tail"
-            backend = Backend \_state _prev _inputs onEvent -> do
-                onEvent
-                    (ToolFinished
-                        (ToolCallResult
-                            "large"
-                            oversized
-                            FunctionCallKind))
-                pure (Left (ConnectionError "down"))
-        config <- testConfig backend
-        execution <-
-            runLoopInputsDetailed config Nothing [UserMessage "hello"]
-        case execution.executionUncommittedDisplayEvents of
-            [ToolFinished result] -> do
-                Text.length result.output
-                    `shouldSatisfy` (<= 2 * 1024 * 1024)
-                result.output `shouldSatisfy`
-                    Text.isPrefixOf "[earlier tool output truncated]"
-                result.output `shouldSatisfy`
-                    Text.isSuffixOf "newest-tail"
-            other ->
-                expectationFailure
-                    ("unexpected display events: " <> show other)
+    FailedDisplay.spec
 
     it "turns synchronous backend exceptions into a failed turn" do
         config <- testConfig $ Backend \_state _prev _inputs _onEvent ->
@@ -2731,111 +2179,6 @@ spec = describe "runLoop" do
 -- Helpers
 --------------------------------------------------------------------------------
 
-testConfig :: Backend -> IO LoopConfig
-testConfig backend = do
-    cancel <- newCancelFlag
-    state <- newIORef emptyBackendSnapshot
-    pure LoopConfig
-        { loopBackend = backend
-        , loopBackendState = BackendStateStore
-            { readBackendState = readIORef state
-            , commitBackendState = \snapshot -> do
-                writeIORef state snapshot
-                pure snapshot
-            }
-        , loopTools = registryFromHandlers
-            [ typedTool "echo" echoArgsDecoder $ \EchoArgs { message } ->
-                pure (Right ("echo:" <> message))
-            ]
-        , loopDispatch = defaultLoopDispatch
-        , loopMaxTurns = defaultLoopMaxTurns
-        , loopOnEvent = \_ -> pure ()
-        , loopApprove = \_ -> pure (Right True)
-        , loopReadSteering = pure []
-        , loopCommitSteering = \_ -> pure ()
-        , loopInterrupt = pure ()
-        , loopCancel = cancel
-        }
-
-registryFromHandlers :: [ToolHandler] -> ToolRegistry
-registryFromHandlers =
-    registryFromPolicies . map (\handler -> (ParallelSafe, handler))
-
-registryFromPolicies :: [(ToolExecutionPolicy, ToolHandler)] -> ToolRegistry
-registryFromPolicies tools =
-    either (error . Text.unpack) id $ mkToolRegistry
-        [ jsonAppToolWithExecution
-            (handlerName handler)
-            ""
-            []
-            AlwaysReadOnly
-            execution
-            handler
-        | (execution, handler) <- tools
-        ]
-
-registryFromTools :: [AppTool] -> ToolRegistry
-registryFromTools =
-    either (error . Text.unpack) id . mkToolRegistry
-
-resourceTool :: Text -> Text -> IO (Either Text Text) -> AppTool
-resourceTool name resource action =
-    withToolResourceClaims
-        (\_ ->
-            pure $ Right
-                [ ToolResourceClaim ToolWrite
-                    (ToolNamedResource resource)
-                ])
-        (jsonAppToolWithExecution
-            name
-            ""
-            []
-            AlwaysReadOnly
-            TurnSequential
-            (noArgsTool name action))
-
-noArgsAppTool :: Text -> IO (Either Text Text) -> AppTool
-noArgsAppTool name action =
-    jsonAppToolWithExecution
-        name
-        ""
-        []
-        AlwaysReadOnly
-        ParallelSafe
-        (noArgsTool name action)
-
-asyncNoArgsTool :: Text -> IO (Either Text Text) -> AppTool
-asyncNoArgsTool name action =
-    withAsyncToolCalls (noArgsAppTool name action)
-
-asyncFunctionToolCall :: Text -> Text -> Text -> ToolCall
-asyncFunctionToolCall callId name arguments =
-    withToolCallMode AsyncToolCall
-        (functionToolCall callId name arguments)
-
-completedCallIds :: [TurnInput] -> [Text]
-completedCallIds inputs =
-    [ result.callId
-    | CompletedTool result <- inputs
-    ]
-
-concurrencyProbeMicros :: Int
-concurrencyProbeMicros = 5000000
-
-data EchoArgs = EchoArgs { message :: Text }
-
-echoArgsDecoder :: Json.Decoder EchoArgs
-echoArgsDecoder = objectArgs $ \object -> EchoArgs <$> reqText object "message"
-
-functionResult :: Text -> Text -> ToolCallResult
-functionResult callId output = ToolCallResultWithOutcome
-    { callId
-    , output
-    , toolResultImages = []
-    , toolResultOutcome = ToolSucceeded
-    , callKind = FunctionCallKind
-    }
-
 completedImageLoopResult :: Either LoopError LoopResult
 completedImageLoopResult =
     Right LoopResult
@@ -3087,47 +2430,3 @@ assertNormalizedDataUrl sourceBytes sourceUrl normalizedUrl = do
                 sourceBytes
                 (Text.drop 5 (Text.dropEnd 7 metadata))
                 normalizedBytes
-
-scriptedBackend
-    :: IORef [(Maybe Text, [TurnInput])]
-    -> [Either ApiError TurnOutput]
-    -> IO Backend
-scriptedBackend submissions answers = do
-    remaining <- newIORef answers
-    pure $ Backend \state prev inputs _onEvent -> do
-        modifyIORef' submissions (++ [(prev, inputs)])
-        atomicModifyIORef' remaining \case
-            [] -> ([], Left (ConnectionError "scripted backend exhausted"))
-            next : rest ->
-                ( rest
-                , fmap
-                    (\output -> BackendResult
-                        { backendOutput = output
-                        , backendState = appendStateMarker state
-                        })
-                    next
-                )
-
-endlessToolsBackend :: IO Backend
-endlessToolsBackend = do
-    counter <- newIORef (0 :: Int)
-    pure $ Backend \state _prev _inputs _onEvent -> do
-        n <- atomicModifyIORef' counter \i -> (i + 1, i + 1)
-        let responseId = "resp-" <> Text.pack (show n)
-        pure $ Right BackendResult
-            { backendOutput = emptyTurnOutput responseId
-                [functionToolCall "c1" "echo" "{\"message\":\"again\"}"]
-                Nothing
-            , backendState = appendStateMarker state
-            }
-
-stateMarker :: ResponseItem
-stateMarker = UnknownResponseItem TaggedObject
-    { tag = "test_state"
-    }
-
-appendStateMarker :: BackendSnapshot -> BackendSnapshot
-appendStateMarker snapshot =
-    advanceBackendSnapshot snapshot
-        (snapshot.backendItems <> [stateMarker])
-        snapshot.backendContinuation
